+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - [Fix] Upgrade marked NPM library to fix a vulnerability (#109)
 - [Fix] Fix dependency packages' vulnerabilities (#124)
 - [Fix] Fix filament display when it has no brand (#125)
+- [Fix] It was possible to delete a material used by filament. It's no more allowed. (#118)
 - [Tec] Upgrade to NodeJS 10 (#107, #122)
 - [Tec] Add `versionSuffix` config setting (#92)
 - [Tec] Improve Docker image and stack (#80 #81 #82 #85 #86 #88)